Update
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@8090 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/mem_pool.h
@@ -113,9 +113,9 @@ int pm_strcpy(POOLMEM *&pm, const char *str);
 int pm_strcpy(POOL_MEM &pm, const char *str);
 int pm_strcpy(POOLMEM *&pm, POOL_MEM &str);
 
-int pm_memcpy(POOLMEM **pm, const char *data, size_t n);
-int pm_memcpy(POOLMEM *&pm, const char *data, size_t n);
-int pm_memcpy(POOL_MEM &pm, const char *data, size_t n);
-int pm_memcpy(POOLMEM *&pm, POOL_MEM &data, size_t n);
+int pm_memcpy(POOLMEM **pm, const char *data, int32_t n);
+int pm_memcpy(POOLMEM *&pm, const char *data, int32_t n);
+int pm_memcpy(POOL_MEM &pm, const char *data, int32_t n);
+int pm_memcpy(POOLMEM *&pm, POOL_MEM &data, int32_t n);
 
 #endif
